src/requirements.txt
adium-theme-ubuntu==0.3.4
argh==0.26.2
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
bcrypt==3.1.3
beautifulsoup4==4.4.1
blinker==1.4
boto==2.48.0
boto3==1.4.5
botocore==1.5.92
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.0.2
cssselect==1.0.1
cycler==0.10.0
Django==1.11.4
django-allauth==0.32.0
django-appconf==1.0.2
django-boto==0.3.12
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-less==0.7.2
django-localflavor==1.5.2
django-pagination==1.0.7
django-pylibmc==0.6.1
django-s3-cache==1.4.2
django-sslserver==0.20
django-static-precompiler==1.6
-e git+https://github.com/jschneier/django-storages-redux.git@dc3902dc226bd69b4698b95c1bf6e6aae9a5fa50#egg=django_storages
django-widget-tweaks==1.4.1
docutils==0.14
dtrx==6.6
enum34==1.1.6
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.2
functools32==3.2.3.post2
futures==3.1.1
html5lib==0.999
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
ipaddr==2.1.11
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
libvirt-python==1.3.1
lxml==3.5.0
MarkupSafe==1.0
matplotlib==2.0.2
ndg-httpsclient==0.4.0
nltk==3.2.4
numpy==1.13.1
oauthlib==2.0.2
olefile==0.44
parsel==1.2.0
pathtools==0.1.2
Pillow==4.1.0
pprint==0.1
psycopg2==2.7.1
pyasn1==0.1.9
pyasn1-modules==0.0.10
pycparser==2.18
pycurl==7.43.0
PyDispatcher==2.0.5
pygal==2.4.0
pygobject==3.20.0
PyJWT==1.5.2
pylibmc==1.5.2
pyOpenSSL==17.2.0
pyparsing==2.2.0
python-apt==1.1.0b1
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
pywatchman==1.3.0
PyYAML==3.12
queuelib==1.4.2
rcssmin==1.0.6
requests==2.18.3
requests-oauthlib==0.8.0
rjsmin==1.0.12
s3transfer==0.1.10
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
SQLAlchemy==1.1.11
sqlparse==0.2.3
stripe==1.53.0
subprocess32==3.2.7
tweepy==3.5.0
Twisted==17.5.0
unity-lens-photos==1.0
urllib3==1.22
vboxapi==1.0
w3lib==1.17.0
watchdog==0.8.3
Werkzeug==0.12.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zope.interface==4.4.2
